refactor: Move random generators to helpers dir
